Make require('config/defer') syntax work
I had presumed this worked before, it is difficult to test this from inside of the project itself.
Manual testing from outside the project tree confirmed that part of the parth after the "/" resolved
relative to the "config" directory where this package is stored, not the lib/ directory where config.js
is stored.