Fix 'vows' export style.
Before, we were giving multiple test suites the same export name of 'DeferredConfig',
which was an accident and could cause problems.

Vows supports a second export style which this commit switches to. The
alternate export style isn't prone to the same copy/paste/modify issue
that the previous style was.

This commits also updates one assert to use a native check for
"isObject".