Resolve defered values in predictable order for consistent results.
Before, we used the `for..in..` construction to iterate over object
properties, which returns the keys in an arbitrary order which is
usually-but-not-always the key insertion order.

Now object keys are traversed in the order returned by calling sort() on
the key names

Having a predictable resolution order could be useful if you want one
defered value to refer to another defered value. Now one should reliably
resolve before the other.