BugFix for case when strict_mode is set and you wish to disable it.
Now we more explicitly define a list of possible values that can be used to enable strict mode.

Before, disabling strict mode by setting the value to "0" would not have worked.

As the node docs say:

"Assigning a property on process.env will implicitly convert the value to a string."
        ref: https://nodejs.org/api/process.html#process_process_env

So, even if you set process.env.NODE_CONFIG_STRICT_MODE=0 this would /not/ disabled
strict me. The zero would get cast to a string, and `if ("0") {...}` would return true.

Likewise, before this change, disabling strict mode by setting the value to "false"
also did not work, as the string "false" is true.

So those cases are fixed. On the other hand, other truthy values like "yes",
"YES" and "TRUE" are no longer accepted.