Fix for #309: Interpret SUPPRESS_NO_CONFIG_WARNING more broadly
This implements the logic described in
https://github.com/lorenwest/node-config/issues/309#issuecomment-212011422

We also let the user know about whether strict mode is currently
on or off for relevant messages.

Additionally, runStrictnessChecks has been renamed to runSanityChecks,
since it is not only run in strict mode, and not all checks are
necessarily affected by strict mode.