Support for nested raw() in javascript configurations
Moving the `raw` wrapped object unwrapping into `makeImmutable`
is piggy-backing on the existing recursive object graph traversal
to ensure that regardless of how deep a raw-wrapped object is
within the configuration, it will be unwrapped.

The ordering of `makeImmutable` and `getImpl` was flipped in order
to consistently return an immutable config. Previously, a mutable
configuration was returned on first `get()`, and an immutable one
on subsequent `get()`s.

Fixes lorenwest/node-config#467