Partial fix for recursion bugs in loadFileConfigs();
Any variables read or written after the first parseFile() call
will likely be incorrect if for any reason loadFileConfigs() ends
up calling itself from an config/*.js files.

The example in the tests illustrates a problem I encountered. This may be
a little off-label, but this same situation could potentially happen with
setModuleDefaults, if it uses loadFileConfigs() to extract the appropriate
values.