Add nyc.
The 'extensions' override here is needed in part because nyc and
the ts-node lazy load code mentioned in #787 fight each other.

Sometimes it's useful to have code coverage in your tests, since
it's a good way to notice dead helper code or stupid things like
accidentally disabling a raft of tests.