Factor out the state tracking in loadFileConfigs.
This is to make space for unravelling the mess that is the singleton
into some functional components that may be useful individually.

This also deprecates some functions that probably should never
have been exposed.

This refactor is as commit-history-preserving as I could manage.
Subsequent commits will involve rearranging more of the code to
extricate things from config.js