Simplify the setModuleDefaults code paths.
Fix a bug where Config.setModuleDefaults after Config.get would
result in the existing Config object not seeing the updates.

This of course only happens when ALLOW_CONFIG_MUTATIONS is true, but
probably worth not introducing a breaking change in 4.0.1

Identified a bug in the configSources handling, but did not fix it.